Expose swagger docs for brig's internal (servantified) endpoints
Start the ToSchema party with brig and cannon

Remove unnecessary module qualifier

Fix type error

Fix

Create instance `ToSchema List1`

Add swagger tags (resulting in sections)

Replace over with %~

Introduce SwaggerTag combinator

Use lens

Add some documentation text

Test RawJson FromJSON/ToJSON

Cleanup

Delete unused endpoint

Align title with cannon's swagger title

Delete Won't-do ToDos

Add changelog

Remove superfluous comment

Add docs about Swagger

Describe the swagger URL pattern better

Use versions in swagger endpoint

Update docs

Move example section

Fix typo

Update swagger.md

Old Swagger is now gone.

Update services/brig/src/Brig/API/Public.hs


Update services/brig/docs/swagger-internal-endpoints.md


Update changelog.d/2-features/internal-endpoints-swagger


Update libs/wire-api/src/Wire/API/SwaggerServant.hs


Update libs/wire-api/test/unit/Test/Wire/API/Roundtrip/Aeson.hs


Remove redundant constraint

Simplify operations tagging with allOperations Traversal

Use LText as data store of RawJson

Replace ToSchema instance with function

Increase lazy-ness of ToSchema RawJson instance

Remove ToSchema for internal data structures

Swagger docs for internal endpoints of Legalhold and Cargohold

Move galley's internal endpoint to wire-api

Swagger for internal Galley endpoints

Add spar's internal endpoints
