+++ libs/wire-api/test/unit/Test/Wire/API/Roundtrip/Aeson.hs
@@ -46,6 +46,7 @@ import qualified Wire.API.Provider.External as Provider.External
 import qualified Wire.API.Provider.Service as Provider.Service
 import qualified Wire.API.Provider.Service.Tag as Provider.Service.Tag
 import qualified Wire.API.Push.Token as Push.Token
+import qualified Wire.API.Routes.Internal.Galley.TeamsIntra as TeamsIntra
 import qualified Wire.API.Team as Team
 import qualified Wire.API.Team.Conversation as Team.Conversation
 import qualified Wire.API.Team.Feature as Team.Feature
@@ -310,7 +311,8 @@ tests =
       testRoundTrip @User.Search.PagingState,
       testRoundTrip @User.Search.TeamContact,
       testRoundTrip @(Wrapped.Wrapped "some_int" Int),
-      testRoundTrip @Conversation.Action.SomeConversationAction
+      testRoundTrip @Conversation.Action.SomeConversationAction,
+      testRoundTrip @TeamsIntra.GuardLegalholdPolicyConflicts
     ]
 
 testRoundTrip ::
