+++ libs/wire-api/wire-api.cabal
@@ -82,8 +82,12 @@ library
     Wire.API.Routes.Internal.Brig.EJPD
     Wire.API.Routes.Internal.Cannon
     Wire.API.Routes.Internal.Cargohold
+    Wire.API.Routes.Internal.Galley
+    Wire.API.Routes.Internal.Galley.ConversationsIntra
     Wire.API.Routes.Internal.Galley.TeamFeatureNoConfigMulti
+    Wire.API.Routes.Internal.Galley.TeamsIntra
     Wire.API.Routes.Internal.LegalHold
+    Wire.API.Routes.Internal.Spar
     Wire.API.Routes.LowLevelStream
     Wire.API.Routes.MultiTablePaging
     Wire.API.Routes.MultiTablePaging.State
