+++ services/galley/src/Galley/Cassandra/Team.hs
@@ -51,11 +51,11 @@ import Galley.Env
 import Galley.Monad
 import Galley.Options
 import Galley.Types.Teams
-import Galley.Types.Teams.Intra
 import Imports hiding (Set, max)
 import Polysemy
 import Polysemy.Input
 import qualified UnliftIO
+import Wire.API.Routes.Internal.Galley.TeamsIntra
 import Wire.API.Team
 import Wire.API.Team.Conversation
 import Wire.API.Team.Member
