+++ services/galley/src/Galley/Effects/TeamStore.hs
@@ -80,12 +80,12 @@ import Data.Id
 import Data.Range
 import Galley.Effects.ListItems
 import Galley.Types.Teams
-import Galley.Types.Teams.Intra
 import Imports
 import Polysemy
 import qualified Proto.TeamEvents as E
 import Wire.API.Error
 import Wire.API.Error.Galley
+import Wire.API.Routes.Internal.Galley.TeamsIntra
 import Wire.API.Team
 import Wire.API.Team.Conversation
 import Wire.API.Team.Member (HardTruncationLimit, TeamMember, TeamMemberList)
