+++ services/galley/test/integration/API.hs
@@ -66,7 +66,6 @@ import Federator.Discovery (DiscoveryFailure (..))
 import Federator.MockServer
 import Galley.API.Mapping
 import Galley.Options (optFederator)
-import Galley.Types.Conversations.Intra
 import Galley.Types.Conversations.Members
 import Imports
 import qualified Network.HTTP.Types.Status as HTTP
@@ -94,6 +93,7 @@ import qualified Wire.API.Federation.API.Galley as F
 import Wire.API.Internal.Notification
 import Wire.API.Message
 import qualified Wire.API.Message as Message
+import Wire.API.Routes.Internal.Galley.ConversationsIntra
 import Wire.API.Routes.MultiTablePaging
 import Wire.API.Routes.Version
 import Wire.API.Routes.Versioned
