+++ tools/stern/src/Stern/Intra.hs
@@ -86,8 +86,6 @@ import Data.Text (strip)
 import Data.Text.Encoding (decodeUtf8, encodeUtf8)
 import Data.Text.Lazy (pack)
 import GHC.TypeLits (KnownSymbol)
-import Galley.Types.Teams.Intra
-import qualified Galley.Types.Teams.Intra as Team
 import Imports
 import Network.HTTP.Types.Method
 import Network.HTTP.Types.Status hiding (statusCode)
@@ -104,6 +102,8 @@ import Wire.API.Internal.Notification
 import Wire.API.Properties
 import Wire.API.Routes.Internal.Brig.Connection
 import qualified Wire.API.Routes.Internal.Brig.EJPD as EJPD
+import Wire.API.Routes.Internal.Galley.TeamsIntra
+import qualified Wire.API.Routes.Internal.Galley.TeamsIntra as Team
 import Wire.API.Routes.Version
 import Wire.API.Routes.Versioned
 import Wire.API.Team
